+++ src/models/post.rs
@@ -1,3 +1,5 @@
+use std::f32::consts::E;
+
 use chrono::{DateTime, Utc};
 use comrak::{
     markdown_to_html_with_plugins,