+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.7] - 2023-03-31
+
+#### Added
+
+- admin api creator post route
+- admin console post creator now is first level functional
+
+#### Fixed
+
+- handle the date better on the updater and creator in the admin console
+
 ## [0.8.6] - 2023-03-30
 
 #### Changed